// korp.qualidade.inspecaoentrada-frontend/libs/quality-qualidade-inspecao-entrada/src/lib/tokens/interfaces/get-planos-inspecao-dto.interface.ts
export interface PlanoInspecaoDTO {
  id: string;
  descricao: string;
  resultado: string;
  maiorValorInspecionado: number;
  menorValorInspecionado: number;
  maiorValorBase: number;
  menorValorBase: number;
  metodo: string;
  observacao: string;
}

export interface GetPlanosInspecaoDTO {
  items: PlanoInspecaoDTO[];
  totalCount: number;
}